Fix incorrect result type on new_buffer_2.
Estimated hours taken: 0.2
Branches: main

library/bit_buffer.m:
	Fix incorrect result type on new_buffer_2.

library/math.m:
	Fix Erlang implementation of math.round/1.

library/profiling_builtin.m:
library/table_builtin.m:
	Add Erlang stubs for :- external procedures for deep profiling and
	tabling.

+++ library/math.m
@@ -456,7 +456,7 @@ math.e = 2.7182818284590452353602874713526625.
     math.round(Num::in) = (Rounded::out),
     [will_not_call_mercury, promise_pure, thread_safe],
 "
-    Rounded = round(Num)
+    Rounded = float(round(Num))
 ").
 math.round(Num) = math.floor(Num + 0.5).
